requirements.txt
matplotlib==3.6.1
numpy==1.22.1
opencv-python==4.6.0.66
Pillow==9.2.0